feat: bookView에서 scss 분리
