+++ src/router/index.ts
@@ -3,7 +3,7 @@ import HomeView from '@/views/HomeView.vue'
 import LoginView from '@/views/LoginView.vue'
 import RegisterView from '@/views/RegisterView.vue'
 import CommunityView from '@/views/CommunityView.vue'
-import BookView from '@/views/BookView.vue'
+import BookView from '@/views/bookView/BookView.vue'
 import LibraryView from '@/views/LibraryView.vue'
 import PerformanceView from '@/views/PerformanceView.vue'
 import ChallengeView from '@/views/ChallengeView.vue'
